refactor: remove empty pages
