+++ frontend/src/app/routes/_authenticated/products/$productId/issues.tsx
@@ -1,6 +1,16 @@
-import { createFileRoute } from "@tanstack/react-router";
-import { ProductIssuesPage } from "@/pages/product-issues";
+import { useSuspenseQuery } from "@tanstack/react-query";
+import { createFileRoute, Link } from "@tanstack/react-router";
+import { useTranslation } from "react-i18next";
 import { getProductByIdQueryOptions } from "@/entities/product";
+import { defaultPagination } from "@/shared/api/schemas.ts";
+import { SetSidebarBreadcrumbs } from "@/shared/components/sidebar-breadcrumbs.tsx";
+import {
+  Breadcrumb,
+  BreadcrumbItem,
+  BreadcrumbLink,
+  BreadcrumbList,
+  BreadcrumbSeparator,
+} from "@/shared/components/ui/breadcrumb.tsx";
 
 export const Route = createFileRoute(
   "/_authenticated/products/$productId/issues",
@@ -14,6 +24,38 @@ export const Route = createFileRoute(
 });
 
 function RouteComponent() {
-  const product = Route.useLoaderData();
-  return <ProductIssuesPage product={product} />;
+  const productId = parseInt(Route.useParams().productId);
+
+  const { data: product } = useSuspenseQuery(
+    getProductByIdQueryOptions(productId),
+  );
+  const { t } = useTranslation("product");
+
+  return (
+    <div>
+      <SetSidebarBreadcrumbs>
+        <Breadcrumb>
+          <BreadcrumbList>
+            <BreadcrumbLink asChild>
+              <Link search={defaultPagination} to="/products">
+                {t("products.label")}
+              </Link>
+            </BreadcrumbLink>
+            <BreadcrumbSeparator />
+            <BreadcrumbItem>
+              <Link
+                to="/products/$productId"
+                params={{ productId: String(product.id) }}
+              >
+                {product.title}
+              </Link>
+            </BreadcrumbItem>
+            <BreadcrumbSeparator />
+            <BreadcrumbItem>{t("items.issues.label")}</BreadcrumbItem>
+          </BreadcrumbList>
+        </Breadcrumb>
+      </SetSidebarBreadcrumbs>
+      <main>Tasks will be here</main>
+    </div>
+  );
 }
