--- frontend/src/pages/product-index/index.tsx
@@ -1,2 +1,0 @@
-// eslint-disable-next-line react-refresh/only-export-components
-export * from "./ui/page";
