--- frontend/src/pages/product-issues/index.ts
@@ -1,1 +1,0 @@
-export * from "./ui/page";
