--- frontend/src/pages/product-issues/ui/page.tsx
@@ -1,56 +1,0 @@
-import { useQuery } from "@tanstack/react-query";
-import { Link } from "@tanstack/react-router";
-import { useTranslation } from "react-i18next";
-import {
-  DetailedProduct,
-  getProductByIdQueryOptions,
-} from "@/entities/product";
-import { defaultPagination } from "@/shared/api/schemas";
-import { SetSidebarBreadcrumbs } from "@/shared/components/sidebar-breadcrumbs.tsx";
-import {
-  Breadcrumb,
-  BreadcrumbItem,
-  BreadcrumbLink,
-  BreadcrumbList,
-  BreadcrumbSeparator,
-} from "@/shared/components/ui/breadcrumb.tsx";
-
-export function ProductIssuesPage({
-  product: initialData,
-}: {
-  product: DetailedProduct;
-}) {
-  const { data: product } = useQuery({
-    ...getProductByIdQueryOptions(initialData.id),
-    initialData,
-  });
-  const { t } = useTranslation("product");
-
-  return (
-    <div>
-      <SetSidebarBreadcrumbs>
-        <Breadcrumb>
-          <BreadcrumbList>
-            <BreadcrumbLink asChild>
-              <Link search={defaultPagination} to="/products">
-                {t("products.label")}
-              </Link>
-            </BreadcrumbLink>
-            <BreadcrumbSeparator />
-            <BreadcrumbItem>
-              <Link
-                to="/products/$productId"
-                params={{ productId: String(product.id) }}
-              >
-                {product.title}
-              </Link>
-            </BreadcrumbItem>
-            <BreadcrumbSeparator />
-            <BreadcrumbItem>{t("items.issues.label")}</BreadcrumbItem>
-          </BreadcrumbList>
-        </Breadcrumb>
-      </SetSidebarBreadcrumbs>
-      <main>Tasks will be here</main>
-    </div>
-  );
-}
