--- frontend/src/pages/product-releases/index.ts
@@ -1,1 +1,0 @@
-export * from "./ui/page";
