--- frontend/src/pages/products-page/index.ts
@@ -1,1 +1,0 @@
-export * from "./ui/page";
