--- frontend/src/pages/profile/index.ts
@@ -1,1 +1,0 @@
-export * from "./page";
